Ability to give away static files from public folder has been added
