+++ package.json
@@ -14,10 +14,13 @@
     "cross-env": "^7.0.3",
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
+    "gravatar": "^1.8.2",
+    "jimp": "^0.16.2",
     "joi": "^17.6.4",
     "jsonwebtoken": "^8.5.1",
     "mongoose": "^6.7.0",
     "morgan": "^1.10.0",
+    "multer": "^1.4.5-lts.1",
     "nanoid": "^3.3.4",
     "passport": "^0.6.0",
     "passport-jwt": "^4.0.0"
@@ -29,6 +32,7 @@
     "eslint-plugin-import": "2.25.3",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-promise": "5.2.0",
+    "jest": "^29.3.1",
     "nodemon": "^2.0.20"
   }
 }
